move page components
