--- src/components/CartItem/index.ts
@@ -1,1 +1,0 @@
-export { CartItemRow } from './CartItemRow';
