+++ src/components/Logo.tsx
@@ -1,7 +1,7 @@
 import { Anchor, Text } from '@mantine/core';
-
 import Link from 'next/link';
-function Logo() {
+
+export function Logo() {
   return (
     <Anchor component={Link} href={{ pathname: '/catalog' }}>
       <Text fz="lg" fw="bolder" tt="uppercase" component="span">
@@ -10,4 +10,3 @@ function Logo() {
     </Anchor>
   );
 }
-export default Logo;
