--- src/components/api.ts
@@ -1,5 +1,0 @@
-function logout() {
-  return fetch('/api/logout', { method: 'post' });
-}
-
-export { logout };
