+++ src/pages/_app.tsx
@@ -1,44 +1,15 @@
-import ColorSchemeToggle from '@components/ColorSchemeToggle';
-import Logo from '@components/Logo';
-import NavigationDrawer from '@components/NavigationDrawer';
+import { Header } from '@components/Header';
 import {
-  ActionIcon,
-  Box,
-  Button,
   ColorScheme,
   ColorSchemeProvider,
   Container,
-  Group,
-  Header as MantineHeader,
-  MantineProvider,
-  MediaQuery,
-  TextInput,
-  ThemeIcon
+  MantineProvider
 } from '@mantine/core';
 import { parse } from 'cookie';
 import useColorSchemeToggle from 'hooks/use-color-scheme-toggle';
 import type { AppContext, AppInitialProps, AppProps } from 'next/app';
 import NextApp from 'next/app';
 import Head from 'next/head';
-import Link from 'next/link';
-import { useRouter } from 'next/router';
-import { BsBoxSeam, BsCart3 } from 'react-icons/bs';
-
-function HiddenMobile({ children }: React.PropsWithChildren) {
-  return (
-    <MediaQuery smallerThan={840} styles={{ display: 'none' }}>
-      {children}
-    </MediaQuery>
-  );
-}
-
-function HiddenDesktop({ children }: React.PropsWithChildren) {
-  return (
-    <MediaQuery largerThan={840} styles={{ display: 'none' }}>
-      {children}
-    </MediaQuery>
-  );
-}
 
 type AppOwnProps = {
   colorScheme: ColorScheme;
@@ -93,69 +64,3 @@ App.getInitialProps = async function getInitialProps(
 };
 
 export default App;
-
-function Header() {
-  const router = useRouter();
-  return (
-    <MantineHeader height={60} px="1rem">
-      <Group position="apart" sx={{ height: '100%' }} spacing={0}>
-        <Logo />
-
-        <HiddenMobile>
-          <Box
-            component="form"
-            method="get"
-            action="/"
-            sx={{ flexGrow: 1, display: 'flex', columnGap: '1rem' }}
-            px="1rem"
-          >
-            <TextInput name="q" sx={{ width: '100%' }} />
-            <Button type="submit">Search</Button>
-          </Box>
-        </HiddenMobile>
-
-        <HiddenMobile>
-          <Group>
-            <ColorSchemeToggle />
-
-            <ActionIcon component={Link} href={{ pathname: '/cart' }}>
-              <ThemeIcon>
-                <BsCart3 />
-              </ThemeIcon>
-            </ActionIcon>
-
-            <ActionIcon component={Link} href={{ pathname: '/orders' }}>
-              <ThemeIcon>
-                <BsBoxSeam />
-              </ThemeIcon>
-            </ActionIcon>
-
-            <Button
-              component={Link}
-              href={{
-                pathname: '/login',
-                query: { callback: router.asPath }
-              }}
-            >
-              Log in
-            </Button>
-            <Button
-              variant="outline"
-              component={Link}
-              href={{
-                pathname: '/signup',
-                query: { callback: router.asPath }
-              }}
-            >
-              Sign up
-            </Button>
-          </Group>
-        </HiddenMobile>
-
-        <HiddenDesktop>
-          <NavigationDrawer />
-        </HiddenDesktop>
-      </Group>
-    </MantineHeader>
-  );
-}
