+++ src/pages/catalog/index.tsx
@@ -1,16 +1,7 @@
-import {
-  Anchor,
-  Box,
-  Button,
-  Card,
-  SimpleGrid,
-  Text,
-  Title
-} from '@mantine/core';
+import { ItemCard } from '@components/catalog/ItemCard';
+import { SimpleGrid, Title } from '@mantine/core';
 import Head from 'next/head';
-import Link from 'next/link';
-import { ListItem, Money } from 'types/catalog';
-import moneyToString from 'utils/money-to-string';
+import { ListItem } from 'types/catalog';
 
 type CatalogPageProps = {
   items?: ListItem[];
@@ -22,7 +13,9 @@ function CatalogPage({ items }: CatalogPageProps) {
       <Head>
         <title>Catalog</title>
       </Head>
+
       <Title>Catalog page</Title>
+
       {items && (
         <SimpleGrid
           breakpoints={[
@@ -37,50 +30,6 @@ function CatalogPage({ items }: CatalogPageProps) {
         </SimpleGrid>
       )}
     </>
-  );
-}
-interface ItemCardProps {
-  item: ListItem;
-}
-function ItemCard({ item: { name, price, uuid } }: ItemCardProps) {
-  return (
-    <Card
-      shadow="sm"
-      p="md"
-      radius="md"
-      withBorder
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'space-between'
-      }}
-    >
-      <Anchor component={Link} href={{ pathname: uuid }} size="md">
-        {name}
-      </Anchor>
-      <Box
-        mt="md"
-        sx={{
-          display: 'flex',
-          justifyContent: 'space-between',
-          alignItems: 'center'
-        }}
-      >
-        <Money>{price}</Money>
-        <Button size="xs">Add to cart</Button>
-      </Box>
-    </Card>
-  );
-}
-interface MoneyProps {
-  children: Money;
-}
-
-function Money({ children }: MoneyProps) {
-  return (
-    <Text weight="bold" size="xl" component="span">
-      {moneyToString(children)}
-    </Text>
   );
 }
